Move export pdf button to footer
